requirements.txt
supabase==2.17.0
openai==1.75.0
streamlit==1.45.1
python-dotenv==1.1.0
bcrypt==4.1.2